auth setup

+++ backend/controllers/goalController.js
@@ -1,9 +1,10 @@
 const expressAsyncHandler = require('express-async-handler')
 const asynHandler = require('express-async-handler')
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
 
 const getGoals = asynHandler(async(req, res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({ user: req.user.id })
     res.status(200).json(goals)
 })
 
@@ -13,6 +14,7 @@ const addGoals = asynHandler( async(req, res)=>{
         throw new Error('Please add a text field')
     }
     const goal = await Goal.create({
+        user: req.user.id,
         text: req.body.text
     })
     res.status(200).json(goal)
@@ -24,6 +26,19 @@ const updateGoals = asynHandler(async(req, res)=>{
         res.status(400)
         throw new Error('Goal not Found')
     }
+
+    const user = await User.findById(req.user.id)
+
+    if(!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    if(goal.user.toString() !== req.user.id) {
+        res.status(401)
+        throw new Error('unauthorized user')
+    }
+
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {new: true})
     res.status(200).json({ message: `update goal ${req.params.id}` })
 }
@@ -34,6 +49,19 @@ const deleteGoals =  asynHandler(async(req, res)=>{
         res.status(400)
         throw new Error('Goal not Found')
     }
+    
+    const user = await User.findById(req.user.id)
+
+    if(!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    if(goal.user.toString() !== req.user.id) {
+        res.status(401)
+        throw new Error('unauthorized user')
+    }
+
     await Goal.deleteOne({id: req.params.id})
     res.status(200).json(req.params.id)
 }
